Refactor asset drawer state and add API loader
