+++ server/src/components/assets/AssetDetailDrawer.tsx
@@ -1,718 +1,121 @@
-'use client';
-
-import { useEffect, useMemo, useRef, useState } from 'react';
-import { useRouter } from 'next/navigation';
-import Drawer from 'server/src/components/ui/Drawer';
-import { Tabs, TabsList, TabsTrigger, TabsContent } from 'server/src/components/ui/Tabs';
-import { Badge } from 'server/src/components/ui/Badge';
-import { Button } from 'server/src/components/ui/Button';
-import { Card } from 'server/src/components/ui/Card';
-import {
+import type {
   Asset,
   AssetHistory,
   AssetMaintenanceHistory,
   AssetMaintenanceReport,
   AssetTicketSummary,
-  isMobileDeviceAsset,
-  isNetworkDeviceAsset,
-  isPrinterAsset,
-  isServerAsset,
-  isWorkstationAsset
 } from 'server/src/interfaces/asset.interfaces';
-import { IDocument } from 'server/src/interfaces/document.interface';
-import { getAssetDetailBundle } from 'server/src/lib/actions/asset-actions/assetActions';
-import type { AssetDetailBundle } from 'server/src/lib/actions/asset-actions/assetActions';
-import { withDataAutomationId } from 'server/src/types/ui-reflection/withDataAutomationId';
-import { useRegisterUIComponent } from 'server/src/types/ui-reflection/useRegisterUIComponent';
-import type { ContainerComponent } from 'server/src/types/ui-reflection/types';
-import { Clock3, Copy, FileText, Layers, Link as LinkIcon, ListChecks, Settings2, ShieldCheck } from 'lucide-react';
-import AssetDocuments from './AssetDocuments';
-import CreateTicketFromAssetButton from './CreateTicketFromAssetButton';
-import DeleteAssetButton from './DeleteAssetButton';
+import type { IDocument } from 'server/src/interfaces/document.interface';
+import {
+  getAsset,
+  getAssetHistory,
+  getAssetMaintenanceReport,
+  getAssetLinkedTickets,
+} from 'server/src/lib/actions/asset-actions/assetActions';
+import { getAssetDocuments } from 'server/src/lib/actions/asset-actions/assetDocumentActions';
+import {
+  ASSET_DRAWER_TABS,
+  panelParamToTab,
+  type AssetDetailDrawerProps,
+  type AssetDrawerServerData,
+  type AssetDrawerTab,
+} from './AssetDetailDrawer.types';
 
-export const ASSET_DRAWER_TABS = {
-  OVERVIEW: 'Overview',
-  MAINTENANCE: 'Maintenance',
-  TICKETS: 'Tickets',
-  CONFIGURATION: 'Configuration',
-  DOCUMENTS: 'Documents'
-} as const;
+export { ASSET_DRAWER_TABS };
+export type { AssetDrawerTab };
 
-export type AssetDrawerTab = typeof ASSET_DRAWER_TABS[keyof typeof ASSET_DRAWER_TABS];
-
-interface AssetDetailDrawerProps {
-  assetId: string | null;
-  isOpen: boolean;
-  activeTab: AssetDrawerTab;
-  onTabChange: (tab: AssetDrawerTab) => void;
-  onClose: () => void;
-}
-
-const TAB_ORDER: AssetDrawerTab[] = [
-  ASSET_DRAWER_TABS.OVERVIEW,
-  ASSET_DRAWER_TABS.MAINTENANCE,
-  ASSET_DRAWER_TABS.TICKETS,
-  ASSET_DRAWER_TABS.CONFIGURATION,
-  ASSET_DRAWER_TABS.DOCUMENTS
-];
-
-const STATUS_STYLES: Record<string, string> = {
-  active: 'bg-emerald-100 text-emerald-700 ring-1 ring-emerald-600/20',
-  inactive: 'bg-gray-100 text-gray-700 ring-1 ring-gray-600/20',
-  maintenance: 'bg-amber-100 text-amber-700 ring-1 ring-amber-600/20'
-};
-
-export function AssetDetailDrawer({ assetId, isOpen, activeTab, onTabChange, onClose }: AssetDetailDrawerProps) {
-  const router = useRouter();
-  const bundleCacheRef = useRef<Map<string, AssetDetailBundle>>(new Map());
-  const [currentBundle, setCurrentBundle] = useState<AssetDetailBundle | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const registerDrawer = useRegisterUIComponent<ContainerComponent>({
-    id: 'asset-detail-drawer',
-    type: 'container',
-    label: 'Asset Detail Drawer'
-  });
-
-
-  useEffect(() => {
-    if (!assetId || !isOpen) {
-      setCurrentBundle(null);
-      setIsLoading(false);
-      setError(null);
-      return;
-    }
-
-    let isCancelled = false;
-    const cached = bundleCacheRef.current.get(assetId);
-    if (cached) {
-      setCurrentBundle(cached);
-      setIsLoading(false);
-      setError(null);
-    } else {
-      setCurrentBundle(null);
-      setIsLoading(true);
-      setError(null);
-    }
-
-    const loadBundle = async () => {
-      try {
-        const bundle = await getAssetDetailBundle(assetId);
-        if (isCancelled) {
-          return;
-        }
-        bundleCacheRef.current.set(assetId, bundle);
-        setCurrentBundle(bundle);
-        setIsLoading(false);
-        setError(null);
-      } catch (err) {
-        console.error('Failed to load asset drawer data:', err);
-        if (isCancelled) {
-          return;
-        }
-        setIsLoading(false);
-        setError('Unable to load asset details right now. Please try again.');
-        if (!cached) {
-          setCurrentBundle(null);
-        }
-      }
-    };
-
-    if (!cached) {
-      void loadBundle();
-    }
-
-    return () => {
-      isCancelled = true;
-    };
-  }, [assetId, isOpen]);
-
-  const asset = currentBundle?.asset ?? null;
-  const maintenanceReport = currentBundle?.maintenanceReport ?? null;
-  const maintenanceHistory = currentBundle?.maintenanceHistory ?? [];
-  const history = currentBundle?.history ?? [];
-  const tickets = currentBundle?.tickets ?? [];
-  const documents = currentBundle?.documents ?? [];
-
-  useEffect(() => {
-    const asset = currentBundle?.asset;
-    if (!assetId) {
-      registerDrawer?.({
-        helperText: 'Awaiting asset selection'
-      });
-    } else if (asset) {
-      registerDrawer?.({
-        helperText: `${asset.name} • ${activeTab}`
-      });
-    }
-  }, [assetId, activeTab, currentBundle, registerDrawer]);
-
-  const statusBadge = useMemo(() => {
-    if (!asset) {
-      return null;
-    }
-    const statusClass = STATUS_STYLES[asset.status] || 'bg-blue-100 text-blue-700 ring-1 ring-blue-600/20';
-    return (
-      <Badge
-        id="asset-drawer-status"
-        variant="outline"
-        className={`px-2 py-1 text-xs font-semibold ${statusClass}`}
-      >
-        {asset.status.charAt(0).toUpperCase() + asset.status.slice(1)}
-      </Badge>
-    );
-  }, [asset]);
-
-  const overviewTab = useMemo(() => {
-    if (!asset) {
-      return renderEmptyState('Select an asset to view details');
-    }
-
-    return (
-      <div className="space-y-6">
-        <div className="flex flex-col gap-3 md:flex-row md:items-start md:justify-between">
-          <div className="space-y-1">
-            <div className="flex items-center gap-3">
-              <h2 className="text-2xl font-semibold text-gray-900">{asset.name}</h2>
-              {statusBadge}
-            </div>
-            <p className="text-sm text-gray-500">
-              Asset tag {asset.asset_tag} &bull; {asset.asset_type.replace('_', ' ')}
-            </p>
-            {asset.client?.client_name && (
-              <p className="text-sm text-gray-500">Client: {asset.client.client_name}</p>
-            )}
-          </div>
-          <div className="flex flex-wrap gap-2">
-            <Button
-              id="asset-drawer-open-record"
-              variant="secondary"
-              size="sm"
-              className="gap-2"
-              onClick={() => router.push(`/msp/assets/${asset.asset_id}`)}
-            >
-              <FileText className="h-4 w-4" /> Open asset record
-            </Button>
-            <CreateTicketFromAssetButton asset={asset} />
-            <DeleteAssetButton
-              assetId={asset.asset_id}
-              assetName={asset.name}
-              variant="ghost"
-              size="sm"
-              label="Delete"
-              onDeleted={() => {
-                onClose();
-              }}
-            />
-          </div>
-        </div>
-
-        {maintenanceReport ? (
-          <div className="grid grid-cols-1 sm:grid-cols-2 gap-4" {...withDataAutomationId({ id: 'asset-drawer-overview-metrics' })}>
-            <MetricCard
-              id="metric-maintenance-total"
-              icon={<ListChecks className="h-4 w-4 text-blue-500" />}
-              label="Active Schedules"
-              value={maintenanceReport.active_schedules}
-            />
-            <MetricCard
-              id="metric-maintenance-upcoming"
-              icon={<Clock3 className="h-4 w-4 text-amber-500" />}
-              label="Upcoming Maintenance"
-              value={maintenanceReport.upcoming_maintenances}
-              helper={maintenanceReport.next_maintenance ? `Next on ${formatDate(maintenanceReport.next_maintenance)}` : undefined}
-            />
-            <MetricCard
-              id="metric-maintenance-completed"
-              icon={<ShieldCheck className="h-4 w-4 text-emerald-500" />}
-              label="Completed Maintenance"
-              value={maintenanceReport.completed_maintenances}
-              helper={`Compliance ${Math.round(maintenanceReport.compliance_rate)}%`}
-            />
-            <MetricCard
-              id="metric-maintenance-last"
-              icon={<Layers className="h-4 w-4 text-violet-500" />}
-              label="Last Maintenance"
-              value={maintenanceReport.last_maintenance ? formatRelative(maintenanceReport.last_maintenance) : 'No history'}
-            />
-          </div>
-        ) : (
-          <div className="flex items-center gap-3 rounded-lg border border-dashed border-gray-200 p-4 text-sm text-gray-500">
-            <Clock3 className="h-4 w-4 text-gray-400" />
-            Maintenance data will appear once schedules are configured.
-          </div>
-        )}
-
-        <Card className="space-y-4 p-4" {...withDataAutomationId({ id: 'asset-drawer-overview-info' })}>
-          <SectionTitle icon={<Settings2 className="h-4 w-4" />} title="Asset summary" />
-          <InfoGrid asset={asset} />
-        </Card>
-
-        {history && history.length > 0 && (
-          <Card className="space-y-4 p-4" {...withDataAutomationId({ id: 'asset-drawer-overview-history' })}>
-            <SectionTitle icon={<Copy className="h-4 w-4" />} title="Recent lifecycle events" />
-            <ul className="space-y-3">
-              {history.slice(0, 5).map(event => (
-                <li key={event.history_id} className="flex items-start gap-3">
-                  <div className="h-2 w-2 translate-y-2 rounded-full bg-primary-500" />
-                  <div className="space-y-1">
-                    <p className="text-sm font-medium text-gray-900 capitalize">{event.change_type.replace('_', ' ')}</p>
-                    <p className="text-xs text-gray-500">{formatRelative(event.changed_at)}</p>
-                  </div>
-                </li>
-              ))}
-            </ul>
-          </Card>
-        )}
-      </div>
-    );
-  }, [asset, history, maintenanceReport, router, statusBadge, onClose]);
-
-  const maintenanceTab = useMemo(() => {
-    if (!asset) {
-      return renderEmptyState('Select an asset to review maintenance activity');
-    }
-
-    if (!maintenanceReport) {
-      return renderEmptyState('No maintenance schedules found for this asset yet.');
-    }
-
-    return (
-      <div className="space-y-6">
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-          <MetricCard
-            id="metric-schedules-total"
-            icon={<ListChecks className="h-4 w-4 text-blue-500" />}
-            label="Total Schedules"
-            value={maintenanceReport.total_schedules}
-          />
-          <MetricCard
-            id="metric-schedules-active"
-            icon={<ShieldCheck className="h-4 w-4 text-emerald-500" />}
-            label="Active"
-            value={maintenanceReport.active_schedules}
-          />
-          <MetricCard
-            id="metric-schedules-compliance"
-            icon={<Layers className="h-4 w-4 text-violet-500" />}
-            label="Compliance Rate"
-            value={`${Math.round(maintenanceReport.compliance_rate)}%`}
-          />
-        </div>
-
-        {maintenanceHistory && maintenanceHistory.length > 0 ? (
-          <Card className="space-y-4 p-4" {...withDataAutomationId({ id: 'asset-drawer-maintenance-history' })}>
-            <SectionTitle icon={<Clock3 className="h-4 w-4" />} title="Maintenance history" />
-            <ul className="space-y-4">
-              {maintenanceHistory.map(entry => (
-                <li key={entry.history_id} className="flex items-start gap-3">
-                  <div className="mt-1 h-2 w-2 rounded-full bg-primary-500" />
-                  <div>
-                    <p className="text-sm font-medium text-gray-900">
-                      {formatDate(entry.performed_at)}
-                    </p>
-                    <p className="text-xs text-gray-500">
-                      Logged by {entry.performed_by || 'system'}
-                    </p>
-                    {entry.notes && (
-                      <p className="mt-1 text-sm text-gray-600">{entry.notes}</p>
-                    )}
-                  </div>
-                </li>
-              ))}
-            </ul>
-          </Card>
-        ) : (
-          <div className="flex items-center gap-3 rounded-lg border border-dashed border-gray-200 p-4 text-sm text-gray-500">
-            <Clock3 className="h-4 w-4 text-gray-400" />
-            Maintenance logs will appear here after the first service entry.
-          </div>
-        )}
-      </div>
-    );
-  }, [asset, maintenanceHistory, maintenanceReport]);
-
-  const ticketsTab = useMemo(() => {
-    if (!asset) {
-      return renderEmptyState('Select an asset to review linked tickets');
-    }
-
-    if (!tickets || tickets.length === 0) {
-      return renderEmptyState('No tickets linked to this asset yet. Use the quick action to create one.');
-    }
-
-    return (
-      <div className="space-y-3">
-        {tickets.map(ticket => (
-          <Card key={`${ticket.ticket_id}-${ticket.linked_at}`} className="p-4" {...withDataAutomationId({ id: `asset-linked-ticket-${ticket.ticket_id}` })}>
-            <div className="flex flex-col gap-3 md:flex-row md:items-start md:justify-between">
-              <div className="space-y-1">
-                <p className="text-sm font-semibold text-gray-900">{ticket.title}</p>
-                <div className="flex flex-wrap items-center gap-2">
-                  <Badge id={`ticket-status-${ticket.ticket_id}`} variant="outline" className="text-xs">
-                    {ticket.status_name}
-                  </Badge>
-                  {ticket.priority_name && (
-                    <Badge id={`ticket-priority-${ticket.ticket_id}`} variant="secondary" className="text-xs">
-                      {ticket.priority_name}
-                    </Badge>
-                  )}
-                  <span className="text-xs text-gray-500">
-                    Linked {formatRelative(ticket.linked_at)}
-                  </span>
-                </div>
-                <div className="text-xs text-gray-500">
-                  {ticket.client_name && <span>Client: {ticket.client_name}</span>}
-                  {ticket.assigned_to_name && (
-                    <span className="ml-2">Assignee: {ticket.assigned_to_name}</span>
-                  )}
-                </div>
-              </div>
-              <Button
-                id={`asset-ticket-open-${ticket.ticket_id}`}
-                variant="ghost"
-                size="sm"
-                className="gap-2 self-start"
-                onClick={() => router.push(`/msp/tickets/${ticket.ticket_id}`)}
-              >
-                <LinkIcon className="h-4 w-4" /> Open ticket
-              </Button>
-            </div>
-          </Card>
-        ))}
-      </div>
-    );
-  }, [asset, tickets, router]);
-
-  const configurationTab = useMemo(() => {
-    if (!asset) {
-      return renderEmptyState('Select an asset to inspect configuration details');
-    }
-
-    return (
-      <div className="space-y-6">
-        <Card className="space-y-3 p-4" {...withDataAutomationId({ id: 'asset-drawer-configuration-basics' })}>
-          <SectionTitle icon={<Settings2 className="h-4 w-4" />} title="Core attributes" />
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-3 text-sm text-gray-700">
-            <ConfigurationRow label="Serial number" value={asset.serial_number || 'Not provided'} />
-            <ConfigurationRow label="Location" value={asset.location || 'Not provided'} />
-            <ConfigurationRow label="Purchase date" value={asset.purchase_date ? formatDate(asset.purchase_date) : 'Not provided'} />
-            <ConfigurationRow label="Warranty end" value={asset.warranty_end_date ? formatDate(asset.warranty_end_date) : 'Not provided'} />
-          </div>
-        </Card>
-
-        {renderTypeSpecificConfiguration(asset)}
-      </div>
-    );
-  }, [asset]);
-
-  const documentsTab = useMemo(() => {
-    if (!asset) {
-      return renderEmptyState('Select an asset to manage documents');
-    }
-
-    return (
-      <AssetDocuments
-        assetId={asset.asset_id}
-        tenant={asset.tenant}
-        initialDocuments={documents}
-      />
-    );
-  }, [asset, documents]);
-
-  const bundleReady = Boolean(currentBundle);
-  const tabContent = useMemo(() => [
-    { label: ASSET_DRAWER_TABS.OVERVIEW, content: overviewTab },
-    { label: ASSET_DRAWER_TABS.MAINTENANCE, content: maintenanceTab },
-    { label: ASSET_DRAWER_TABS.TICKETS, content: ticketsTab },
-    { label: ASSET_DRAWER_TABS.CONFIGURATION, content: configurationTab },
-    { label: ASSET_DRAWER_TABS.DOCUMENTS, content: documentsTab, forceMount: true }
-  ], [configurationTab, documentsTab, maintenanceTab, overviewTab, ticketsTab]);
-
-  const handleTabChange = (nextTab: string) => {
-    if (TAB_ORDER.includes(nextTab as AssetDrawerTab)) {
-      onTabChange(nextTab as AssetDrawerTab);
-    }
+export async function loadAssetDetailDrawerData({ assetId, panel }: AssetDetailDrawerProps) {
+  const activeTab = panelParamToTab(panel ?? null);
+  const serverData: AssetDrawerServerData = {
+    asset: null,
   };
+  let error: string | null = null;
 
-  const showSkeleton = isLoading && !bundleReady;
+  if (assetId) {
+    try {
+      const asset = await getAsset(assetId);
+      serverData.asset = asset;
 
-  return (
-    <Drawer
-      id="asset-detail-drawer"
-      isOpen={isOpen}
-      onClose={onClose}
-    >
-      <div className="w-[560px] max-w-full space-y-6">
-        <header className="flex items-center justify-between">
-          <div className="space-y-1">
-            <h1 className="text-xl font-semibold text-gray-900">Asset details</h1>
-            <p className="text-sm text-gray-500">
-              Stay in context while reviewing lifecycle and configuration
-            </p>
-          </div>
-        </header>
-
-        {error && (
-          <div className="rounded-lg border border-red-200 bg-red-50 p-4 text-sm text-red-700">
-            {error}
-          </div>
-        )}
-
-        <Tabs
-          value={activeTab}
-          onValueChange={handleTabChange}
-          className="space-y-4"
-        >
-          <TabsList className="w-full gap-2 border-b border-gray-200 text-sm font-medium text-gray-500">
-            {TAB_ORDER.map((tab) => (
-              <TabsTrigger
-                key={tab}
-                value={tab}
-                className="text-sm"
-              >
-                {tab}
-              </TabsTrigger>
-            ))}
-          </TabsList>
-          {tabContent.map(({ label, content, forceMount }) => (
-            <TabsContent
-              key={label}
-              value={label}
-              className="focus:outline-none"
-              forceMount={forceMount}
-            >
-              {showSkeleton ? renderTabSkeleton(label as AssetDrawerTab) : content}
-            </TabsContent>
-          ))}
-        </Tabs>
-      </div>
-    </Drawer>
-  );
-}
-
-function renderTabSkeleton(tab: AssetDrawerTab) {
-  const rows = Array.from({ length: tab === ASSET_DRAWER_TABS.DOCUMENTS ? 6 : 4 });
-  return (
-    <div className="space-y-4">
-      {rows.map((_, index) => (
-        <div
-          key={`${tab}-skeleton-${index}`}
-          className="h-16 rounded-lg bg-gray-100 animate-pulse"
-        />
-      ))}
-    </div>
-  );
-}
-
-function renderEmptyState(message: string) {
-  return (
-    <div className="rounded-lg border border-dashed border-gray-200 p-6 text-center text-sm text-gray-500">
-      {message}
-    </div>
-  );
-}
-
-interface SectionTitleProps {
-  icon: React.ReactNode;
-  title: string;
-}
-
-function SectionTitle({ icon, title }: SectionTitleProps) {
-  return (
-    <div className="flex items-center gap-2 text-sm font-semibold text-gray-900">
-      <span className="rounded-md bg-gray-100 p-1 text-gray-600">
-        {icon}
-      </span>
-      {title}
-    </div>
-  );
-}
-
-interface MetricCardProps {
-  id: string;
-  icon: React.ReactNode;
-  label: string;
-  value: string | number;
-  helper?: string;
-}
-
-function MetricCard({ id, icon, label, value, helper }: MetricCardProps) {
-  return (
-    <Card className="p-4" {...withDataAutomationId({ id })}>
-      <div className="flex items-center gap-3">
-        <div className="rounded-lg bg-gray-100 p-2 text-gray-600">
-          {icon}
-        </div>
-        <div>
-          <p className="text-xs font-medium uppercase tracking-wide text-gray-500">{label}</p>
-          <p className="text-lg font-semibold text-gray-900">{value}</p>
-          {helper && <p className="text-xs text-gray-500">{helper}</p>}
-        </div>
-      </div>
-    </Card>
-  );
-}
-
-function InfoGrid({ asset }: { asset: Asset }) {
-  return (
-    <div className="grid grid-cols-1 md:grid-cols-2 gap-4 text-sm text-gray-600">
-      <InfoRow label="Client" value={asset.client?.client_name || 'Unassigned'} />
-      <InfoRow label="Asset tag" value={asset.asset_tag} />
-      <InfoRow label="Status" value={asset.status} />
-      <InfoRow label="Created" value={formatDate(asset.created_at)} />
-      <InfoRow label="Updated" value={formatRelative(asset.updated_at)} />
-      <InfoRow label="Tenant" value={asset.tenant} />
-    </div>
-  );
-}
-
-interface InfoRowProps {
-  label: string;
-  value: string | number;
-}
-
-function InfoRow({ label, value }: InfoRowProps) {
-  return (
-    <div className="flex flex-col gap-1">
-      <span className="text-xs font-semibold uppercase tracking-wide text-gray-500">{label}</span>
-      <span className="text-sm text-gray-900">{value}</span>
-    </div>
-  );
-}
-
-function ConfigurationRow({ label, value }: InfoRowProps) {
-  return (
-    <div className="flex flex-col">
-      <span className="text-xs font-semibold uppercase tracking-wide text-gray-500">{label}</span>
-      <span className="text-sm text-gray-900">{value}</span>
-    </div>
-  );
-}
-
-function renderTypeSpecificConfiguration(asset: Asset) {
-  switch (asset.asset_type) {
-    case 'workstation':
-      if (asset.workstation && isWorkstationAsset(asset.workstation)) {
-        return (
-          <Card className="space-y-3 p-4" {...withDataAutomationId({ id: 'asset-drawer-config-workstation' })}>
-            <SectionTitle icon={<Settings2 className="h-4 w-4" />} title="Workstation details" />
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-3 text-sm text-gray-700">
-              <ConfigurationRow label="Operating system" value={`${asset.workstation.os_type} ${asset.workstation.os_version}`} />
-              <ConfigurationRow label="CPU" value={`${asset.workstation.cpu_model} (${asset.workstation.cpu_cores} cores)`} />
-              <ConfigurationRow label="RAM" value={`${asset.workstation.ram_gb} GB`} />
-              <ConfigurationRow label="Storage" value={`${asset.workstation.storage_type} • ${asset.workstation.storage_capacity_gb} GB`} />
-              <ConfigurationRow label="GPU" value={asset.workstation.gpu_model || 'Not provided'} />
-              <ConfigurationRow label="Last login" value={asset.workstation.last_login ? formatRelative(asset.workstation.last_login) : 'Never'} />
-            </div>
-          </Card>
-        );
+      if (asset) {
+        await loadTabData(activeTab, asset, serverData);
       }
+    } catch (err) {
+      console.error('Failed to load asset drawer data:', err);
+      error = 'Unable to load asset details right now. Please try again.';
+    }
+  }
+
+  return {
+    activeTab,
+    data: serverData,
+    error,
+  };
+}
+
+async function loadTabData(tab: AssetDrawerTab, asset: Asset, target: AssetDrawerServerData) {
+  switch (tab) {
+    case ASSET_DRAWER_TABS.OVERVIEW: {
+      const [report, history] = await Promise.all([
+        safeGetAssetMaintenanceReport(asset.asset_id),
+        safeGetAssetHistory(asset.asset_id),
+      ]);
+      target.maintenanceReport = report;
+      target.maintenanceHistory = report?.maintenance_history as AssetMaintenanceHistory[] | undefined;
+      target.history = history ?? [];
       break;
-    case 'network_device':
-      if (asset.network_device && isNetworkDeviceAsset(asset.network_device)) {
-        return (
-          <Card className="space-y-3 p-4" {...withDataAutomationId({ id: 'asset-drawer-config-network' })}>
-            <SectionTitle icon={<Settings2 className="h-4 w-4" />} title="Network device details" />
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-3 text-sm text-gray-700">
-              <ConfigurationRow label="Device type" value={asset.network_device.device_type} />
-              <ConfigurationRow label="Management IP" value={asset.network_device.management_ip || 'Not provided'} />
-              <ConfigurationRow label="Port count" value={asset.network_device.port_count} />
-              <ConfigurationRow label="Firmware version" value={asset.network_device.firmware_version} />
-              <ConfigurationRow label="PoE support" value={asset.network_device.supports_poe ? 'Yes' : 'No'} />
-              <ConfigurationRow label="Power draw" value={`${asset.network_device.power_draw_watts} W`} />
-            </div>
-          </Card>
-        );
-      }
+    }
+    case ASSET_DRAWER_TABS.MAINTENANCE: {
+      const report = await safeGetAssetMaintenanceReport(asset.asset_id);
+      target.maintenanceReport = report;
+      target.maintenanceHistory = report?.maintenance_history as AssetMaintenanceHistory[] | undefined;
       break;
-    case 'server':
-      if (asset.server && isServerAsset(asset.server)) {
-        return (
-          <Card className="space-y-3 p-4" {...withDataAutomationId({ id: 'asset-drawer-config-server' })}>
-            <SectionTitle icon={<Settings2 className="h-4 w-4" />} title="Server details" />
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-3 text-sm text-gray-700">
-              <ConfigurationRow label="Operating system" value={`${asset.server.os_type} ${asset.server.os_version}`} />
-              <ConfigurationRow label="CPU" value={`${asset.server.cpu_model} (${asset.server.cpu_cores} cores)`} />
-              <ConfigurationRow label="RAM" value={`${asset.server.ram_gb} GB`} />
-              <ConfigurationRow label="Virtualized" value={asset.server.is_virtual ? 'Yes' : 'No'} />
-              <ConfigurationRow label="Primary IP" value={asset.server.primary_ip || 'Not provided'} />
-              <ConfigurationRow label="Hypervisor" value={asset.server.hypervisor || 'Not provided'} />
-            </div>
-          </Card>
-        );
-      }
+    }
+    case ASSET_DRAWER_TABS.TICKETS: {
+      target.tickets = (await safeGetAssetLinkedTickets(asset.asset_id)) ?? [];
       break;
-    case 'mobile_device':
-      if (asset.mobile_device && isMobileDeviceAsset(asset.mobile_device)) {
-        return (
-          <Card className="space-y-3 p-4" {...withDataAutomationId({ id: 'asset-drawer-config-mobile' })}>
-            <SectionTitle icon={<Settings2 className="h-4 w-4" />} title="Mobile device details" />
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-3 text-sm text-gray-700">
-              <ConfigurationRow label="OS" value={`${asset.mobile_device.os_type} ${asset.mobile_device.os_version}`} />
-              <ConfigurationRow label="Model" value={asset.mobile_device.model} />
-              <ConfigurationRow label="IMEI" value={asset.mobile_device.imei || 'Not provided'} />
-              <ConfigurationRow label="Phone number" value={asset.mobile_device.phone_number || 'Not provided'} />
-              <ConfigurationRow label="Carrier" value={asset.mobile_device.carrier || 'Not provided'} />
-              <ConfigurationRow label="Last check-in" value={asset.mobile_device.last_check_in ? formatRelative(asset.mobile_device.last_check_in) : 'Not reported'} />
-            </div>
-          </Card>
-        );
-      }
+    }
+    case ASSET_DRAWER_TABS.DOCUMENTS: {
+      target.documents = await safeGetAssetDocuments(asset.tenant, asset.asset_id);
       break;
-    case 'printer':
-      if (asset.printer && isPrinterAsset(asset.printer)) {
-        return (
-          <Card className="space-y-3 p-4" {...withDataAutomationId({ id: 'asset-drawer-config-printer' })}>
-            <SectionTitle icon={<Settings2 className="h-4 w-4" />} title="Printer details" />
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-3 text-sm text-gray-700">
-              <ConfigurationRow label="Model" value={asset.printer.model} />
-              <ConfigurationRow label="Network printer" value={asset.printer.is_network_printer ? 'Yes' : 'No'} />
-              <ConfigurationRow label="IP address" value={asset.printer.ip_address || 'Not provided'} />
-              <ConfigurationRow label="Supports color" value={asset.printer.supports_color ? 'Yes' : 'No'} />
-              <ConfigurationRow label="Supports duplex" value={asset.printer.supports_duplex ? 'Yes' : 'No'} />
-              <ConfigurationRow label="Monthly duty cycle" value={asset.printer.monthly_duty_cycle ? `${asset.printer.monthly_duty_cycle} pages` : 'Not provided'} />
-            </div>
-          </Card>
-        );
-      }
-      break;
+    }
+    case ASSET_DRAWER_TABS.CONFIGURATION:
     default:
       break;
   }
-
-  return null;
 }
 
-function formatDate(value: string) {
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return value;
+async function safeGetAssetMaintenanceReport(assetId: string): Promise<AssetMaintenanceReport | null> {
+  try {
+    return await getAssetMaintenanceReport(assetId);
+  } catch (error) {
+    console.error('Failed to load asset maintenance report', error);
+    return null;
   }
-  return date.toLocaleString();
 }
 
-function formatRelative(value: string) {
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return value;
+async function safeGetAssetHistory(assetId: string): Promise<AssetHistory[] | null> {
+  try {
+    return await getAssetHistory(assetId);
+  } catch (error) {
+    console.error('Failed to load asset history', error);
+    return null;
   }
-  const diff = Date.now() - date.getTime();
-  const minutes = Math.floor(diff / (1000 * 60));
-  if (minutes < 1) {
-    return 'moments ago';
+}
+
+async function safeGetAssetLinkedTickets(assetId: string): Promise<AssetTicketSummary[] | null> {
+  try {
+    return await getAssetLinkedTickets(assetId);
+  } catch (error) {
+    console.error('Failed to load linked tickets', error);
+    return null;
   }
-  if (minutes < 60) {
-    return `${minutes} minute${minutes === 1 ? '' : 's'} ago`;
+}
+
+async function safeGetAssetDocuments(tenant: string, assetId: string): Promise<IDocument[] | null> {
+  try {
+    return await getAssetDocuments(tenant, assetId);
+  } catch (error) {
+    console.error('Failed to load asset documents', error);
+    return null;
   }
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) {
-    return `${hours} hour${hours === 1 ? '' : 's'} ago`;
-  }
-  const days = Math.floor(hours / 24);
-  if (days < 7) {
-    return `${days} day${days === 1 ? '' : 's'} ago`;
-  }
-  return date.toLocaleDateString();
 }
